+++ Events/SquadronRankEvent.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EddiDataDefinitions;
+using System;
 using Utilities;
 
 namespace EddiEvents
@@ -13,15 +14,19 @@ namespace EddiEvents
         [PublicAPI("The squadron name")]
         public string name { get; private set; }
 
-        [PublicAPI("The old rank")]
-        public int oldrank { get; private set; }
+        [PublicAPI( "The squadron's numeric ID" )]
+        public int? squadronID { get; private set; }
 
-        [PublicAPI("The new rank")]
-        public int newrank { get; private set; }
+        [PublicAPI( "Your old squadron rank, as an object with properties 'rankID', 'invariantName', and 'localizedName'" )]
+        public SquadronRank oldrank { get; private set; }
 
-        public SquadronRankEvent(DateTime timestamp, string name, int oldrank, int newrank) : base(timestamp, NAME)
+        [PublicAPI( "Your new squadron rank, as an object with properties 'rankID', 'invariantName', and 'localizedName'" )]
+        public SquadronRank newrank { get; private set; }
+
+        public SquadronRankEvent ( DateTime timestamp, string name, int? squadronID, SquadronRank oldrank, SquadronRank newrank ) : base(timestamp, NAME)
         {
             this.name = name;
+            this.squadronID = squadronID;
             this.oldrank = oldrank;
             this.newrank = newrank;
         }